+++ src/main/java/org/betonquest/betonquest/VariableString.java
@@ -15,6 +15,11 @@ import java.util.regex.Pattern;
  * Makes handling instructions with variables easier.
  */
 public class VariableString {
+    /**
+     * The pattern to match variables in a string marked with percent signs.<br>
+     * The percentage can be escaped with a backslash, and the backslash can be escaped with another backslash.
+     */
+    private static final Pattern VARIABLE_PATTERN = Pattern.compile("(?<!\\\\)(?:\\\\\\\\)*(%((?:[^%\\\\]|\\\\.)*?)%)(?<!\\\\)(?:\\\\\\\\)*");
 
     /**
      * The string that may contain variables.
@@ -64,7 +69,7 @@ public class VariableString {
 
         for (final String variable : resolveVariables(this.string)) {
             try {
-                BetonQuest.createVariable(questPackage, variable);
+                BetonQuest.createVariable(questPackage, replaceEscapedPercent(variable));
             } catch (final InstructionParseException exception) {
                 throw new InstructionParseException("Could not create '" + variable + "' variable: "
                         + exception.getMessage(), exception);
@@ -77,7 +82,7 @@ public class VariableString {
 
     private List<String> resolveVariables(final String text) {
         final List<String> variables = new ArrayList<>();
-        final Matcher matcher = Pattern.compile("%[^ %\\s]+%").matcher(text);
+        final Matcher matcher = VARIABLE_PATTERN.matcher(text);
         while (matcher.find()) {
             final String variable = matcher.group();
             if (!variables.contains(variable)) {
@@ -96,10 +101,14 @@ public class VariableString {
     public String getString(@Nullable final Profile profile) {
         String resolvedString = string;
         for (final String variable : variables) {
-            final String resolvedVariable = BetonQuest.getInstance().getVariableValue(questPackage.getQuestPath(), variable, profile);
+            final String resolvedVariable = BetonQuest.getInstance().getVariableValue(questPackage.getQuestPath(), replaceEscapedPercent(variable), profile);
             resolvedString = resolvedString.replace(variable, resolvedVariable);
         }
         return resolvedString;
+    }
+
+    private String replaceEscapedPercent(final String input) {
+        return input.replaceAll("(?<!\\\\)\\\\%", "%");
     }
 
     /**