+++ CHANGELOG.md
@@ -22,6 +22,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### Fixed
 - NPE if a player replies to a SlowTellRaw conversation IO when the text is not completely displayed
 - head `QuestItem` matches even if they are distinct 
+- modulo operator in math variable can now be used by escaping the percent sign
 ### Security
 
 ## [2.1.1] - 2024-05-09